pyproject.toml
[tool.poetry]
name = "drf-online-school"
version = "0.1.0"
description = ""
authors = ["Евгений Ермак <144560375+EvgeniiErmak@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
djangorestframework = "^3.14.0"
pillow = "^10.2.0"
psycopg2-binary = "^2.9.9"
django-extensions = "^3.2.3"
ipython = "^8.21.0"
django-template-bootstrap = "^0.2.15"
django-bootstrap4 = "^24.1"
django-bootstrap5 = "^23.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
